# internal/wire/schema/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "LeX API",
    "version": "0.1-draft",
    "summary": "Synchronous validation and caller-owned replay. This document is a working draft, not a certification."
  },
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "health",
        "responses": {
          "200": {"description": "Process is serving."}
        }
      }
    },
    "/v1/capabilities": {
      "get": {
        "operationId": "capabilities",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Supported operations. Evaluation follows adapter configuration. Execution is off.",
            "headers": {"Cache-Control": {"schema": {"const": "no-store"}}},
            "content": {
              "application/json": {"schema": {"$ref": "#/components/schemas/CapabilitiesResponse"}}
            }
          },
          "400": {"$ref": "#/components/responses/Problem"},
          "401": {"$ref": "#/components/responses/Problem"},
          "403": {"$ref": "#/components/responses/Problem"},
          "405": {"$ref": "#/components/responses/Problem"},
          "406": {"$ref": "#/components/responses/Problem"},
          "413": {"$ref": "#/components/responses/Problem"},
          "503": {"$ref": "#/components/responses/Problem"}
        }
      }
    },
    "/v1/evaluations": {
      "post": {
        "operationId": "evaluate",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {"schema": {"$ref": "#/components/schemas/EvaluationRequest"}}
          }
        },
        "responses": {
          "200": {
            "description": "Synchronous non-error verdict: validated, rejected, insufficient, conflict, or manual_review. A technical error verdict is not HTTP 200. No Location and no 202 job.",
            "headers": {"Cache-Control": {"schema": {"const": "no-store"}}},
            "content": {
              "application/json": {"schema": {"$ref": "#/components/schemas/EvaluationResponse"}}
            }
          },
          "400": {"$ref": "#/components/responses/Problem"},
          "401": {"$ref": "#/components/responses/Problem"},
          "403": {"$ref": "#/components/responses/Problem"},
          "405": {"$ref": "#/components/responses/Problem"},
          "406": {"$ref": "#/components/responses/Problem"},
          "413": {"$ref": "#/components/responses/Problem"},
          "415": {"$ref": "#/components/responses/Problem"},
          "422": {"$ref": "#/components/responses/Problem"},
          "499": {"$ref": "#/components/responses/Problem"},
          "500": {"$ref": "#/components/responses/Problem"},
          "502": {"$ref": "#/components/responses/Problem"},
          "503": {"$ref": "#/components/responses/Problem"},
          "504": {"$ref": "#/components/responses/Problem"}
        }
      }
    },
    "/v1/replays": {
      "post": {
        "operationId": "replay",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {"schema": {"$ref": "replay-bundle.schema.json"}}
          }
        },
        "responses": {
          "200": {
            "description": "Deterministic verdict from the submitted bundle, including a reproduced technical error verdict.",
            "headers": {"Cache-Control": {"schema": {"const": "no-store"}}},
            "content": {
              "application/json": {"schema": {"$ref": "#/components/schemas/ReplayResponse"}}
            }
          },
          "400": {"$ref": "#/components/responses/Problem"},
          "401": {"$ref": "#/components/responses/Problem"},
          "403": {"$ref": "#/components/responses/Problem"},
          "405": {"$ref": "#/components/responses/Problem"},
          "406": {"$ref": "#/components/responses/Problem"},
          "413": {"$ref": "#/components/responses/Problem"},
          "415": {"$ref": "#/components/responses/Problem"},
          "422": {"$ref": "#/components/responses/Problem"},
          "499": {"$ref": "#/components/responses/Problem"},
          "500": {"$ref": "#/components/responses/Problem"},
          "503": {"$ref": "#/components/responses/Problem"},
          "504": {"$ref": "#/components/responses/Problem"}
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer"}
    },
    "responses": {
      "Problem": {
        "description": "RFC 9457 problem. The status member matches the HTTP status.",
        "content": {
          "application/problem+json": {
            "schema": {"$ref": "#/components/schemas/Problem"},
            "example": {
              "type": "https://lex.fastygo.dev/problems/authentication_required",
              "title": "Unauthorized",
              "status": 401,
              "detail": "a bearer token is required",
              "reason": "authentication_required"
            }
          }
        }
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "title", "status", "detail", "reason"],
        "properties": {
          "type": {"type": "string"},
          "title": {"type": "string"},
          "status": {"type": "integer"},
          "detail": {"type": "string"},
          "reason": {"type": "string", "enum": [
            "admission_limited",
            "authentication_denied",
            "authentication_required",
            "body_too_large",
            "client_canceled",
            "deadline_exceeded",
            "decision_error",
            "decision_provider_unavailable",
            "invalid_body",
            "invalid_json",
            "invalid_replay_bundle",
            "internal_error",
            "json_too_deep",
            "method_not_allowed",
            "not_acceptable",
            "not_found",
            "pack_error",
            "policy_error",
            "project_forbidden",
            "provider_unavailable",
            "question_error",
            "response_budget",
            "unsupported_media_type",
            "verification_error"
          ]},
          "verdict": {"type": "string", "enum": ["validated", "rejected", "insufficient", "conflict", "manual_review", "error"]},
          "findings": {"type": "array", "items": {"$ref": "#/components/schemas/Finding"}},
          "trace": {"type": "array"},
          "replay_bundle": {"type": "object"},
          "adapter_metadata": {"$ref": "#/components/schemas/AdapterMetadata"}
        }
      },
      "CapabilitiesResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["protocol_status", "context", "operations", "policy", "limits", "retention"],
        "properties": {
          "protocol_status": {"const": "working_draft"},
          "context": {
            "type": "object",
            "additionalProperties": false,
            "required": ["capability", "retrieval"],
            "properties": {
              "capability": {"const": "memory-exact-v1"},
              "retrieval": {"const": "exact_phrase"}
            }
          },
          "operations": {
            "type": "object",
            "additionalProperties": false,
            "required": ["evaluation", "replay", "execution"],
            "properties": {
              "evaluation": {"type": "boolean"},
              "replay": {"const": true},
              "execution": {"const": false}
            }
          },
          "policy": {"$ref": "#/components/schemas/PolicyDisclosure"},
          "limits": {
            "type": "object",
            "additionalProperties": false,
            "description": "process_admission is the maximum number of requests running in this process, not a global quota.",
            "required": ["max_sources", "max_body_bytes", "focus_max_items", "focus_max_chars", "request_timeout_ms", "process_admission"],
            "properties": {
              "max_sources": {"const": 128},
              "max_body_bytes": {"type": "integer", "minimum": 1, "maximum": 2097152},
              "focus_max_items": {"const": 8},
              "focus_max_chars": {"const": 65536},
              "request_timeout_ms": {"type": "integer", "minimum": 1, "maximum": 20000},
              "process_admission": {"type": "integer", "minimum": 1, "maximum": 4}
            }
          },
          "retention": {"$ref": "#/components/schemas/Retention"}
        }
      },
      "EvaluationResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["protocol_version", "verdict", "findings", "context_runtime", "replay_available", "trace", "policy", "retention"],
        "properties": {
          "protocol_version": {"const": "0.1-draft"},
          "verdict": {"enum": ["validated", "rejected", "insufficient", "conflict", "manual_review"]},
          "findings": {"type": "array", "items": {"$ref": "#/components/schemas/Finding"}},
          "context_runtime": {"const": "memory-exact-v1"},
          "resolved_model": {"type": "string", "minLength": 1},
          "replay_available": {"type": "boolean"},
          "replay_bundle": {"type": "object"},
          "stage": {"type": "string"},
          "trace": {"type": "array", "items": {"$ref": "#/components/schemas/TraceStage"}},
          "policy": {"$ref": "#/components/schemas/PolicyDisclosure"},
          "retention": {"$ref": "#/components/schemas/Retention"},
          "adapter_metadata": {"$ref": "#/components/schemas/AdapterMetadata"}
        }
      },
      "ReplayResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["replay_status", "verdict", "findings", "policy_calibration", "retention", "trace"],
        "properties": {
          "replay_status": {"const": "verdict_reproduced"},
          "verdict": {"$ref": "#/components/schemas/Verdict"},
          "findings": {"type": "array", "items": {"$ref": "#/components/schemas/Finding"}},
          "policy_calibration": {"const": "uncalibrated"},
          "retention": {"$ref": "#/components/schemas/Retention"},
          "trace": {"type": "array", "items": {"$ref": "#/components/schemas/TraceStage"}}
        }
      },
      "Finding": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "verdict", "detail"],
        "properties": {
          "code": {"type": "string", "pattern": "^(action_inconsistent|binding_mismatch|checksum_mismatch|chunk_identity|entity_checksum_mismatch|establishment_below_threshold|evidence_conflict|inference_only|instruction_in_evidence|invalid_answers|missing_provenance|negative_result|no_eligible_evidence|pack_envelope|pack_rebuild|pack_request_identity|pack_shape|partial_surface|project_binding|query_mismatch|rejection_mismatch|review_required|safety_gate|snapshot_identity|source_admission|support_below_threshold|unknown_adapter|unpinned_adapter|unpinned_entity|unpinned_focus|unpinned_policy|unpinned_question_set|unpinned_verifier|unresolved_model|(answer_type_mismatch|invalid_choice|invalid_noul|invalid_score|missing_answer|unexpected_answer|unsupported_question_type):.+)$"},
          "verdict": {"$ref": "#/components/schemas/Verdict"},
          "detail": {"type": "string", "minLength": 1}
        }
      },
      "AdapterMetadata": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "request_id": {"type": "string", "minLength": 1, "maxLength": 256},
          "evaluation_time_ms": {"type": "number", "minimum": 0},
          "usage": {
            "type": "object",
            "minProperties": 1,
            "additionalProperties": {"type": "number", "minimum": 0}
          }
        }
      },
      "TraceStage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "status"],
        "properties": {
          "name": {"enum": ["receive", "pack", "decide", "verify", "replay"]},
          "status": {"enum": ["completed", "failed", "skipped"]}
        }
      },
      "PolicyDisclosure": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "version", "calibration"],
        "properties": {
          "id": {"const": "claim-validation"},
          "version": {"const": "0.1.0"},
          "calibration": {"const": "uncalibrated"}
        }
      },
      "Retention": {
        "type": "object",
        "additionalProperties": false,
        "required": ["server_history", "replay", "idempotency"],
        "properties": {
          "server_history": {"const": false},
          "replay": {"const": "caller_owned"},
          "idempotency": {"const": "none"}
        }
      },
      "Verdict": {
        "type": "string",
        "enum": ["validated", "rejected", "insufficient", "conflict", "manual_review", "error"]
      },
      "EvaluationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["project_id", "entity", "query", "sources"],
        "properties": {
          "project_id": {"type": "string", "minLength": 1, "maxLength": 256},
          "query": {"type": "string", "minLength": 1, "maxLength": 4096},
          "metadata": {
            "type": "object",
            "maxProperties": 8,
            "propertyNames": {"pattern": "^[a-z][a-z0-9_]{0,63}$"},
            "additionalProperties": {"type": "string", "maxLength": 256}
          },
          "entity": {
            "type": "object",
            "additionalProperties": false,
            "required": ["id", "type", "schema_version", "version"],
            "properties": {
              "id": {"type": "string", "minLength": 1, "maxLength": 256},
              "type": {"const": "claim"},
              "schema_version": {"const": "0.1"},
              "version": {"type": "string", "minLength": 1, "maxLength": 256}
            }
          },
          "sources": {
            "type": "array",
            "minItems": 1,
            "maxItems": 128,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": ["id", "version", "text"],
              "properties": {
                "id": {"type": "string", "minLength": 1, "maxLength": 256},
                "version": {"type": "string", "minLength": 1, "maxLength": 256},
                "text": {"type": "string", "minLength": 1, "maxLength": 262144}
              }
            }
          }
        }
      }
    }
  }
}